Added str helpers, switched hours_used to long

+++ src/Classes/ProcRecord.cs
@@ -3,7 +3,7 @@
         public int Id { get; set; }
         public string proc_name { get; set; }
         public string program_name { get; set; }
-        public int hours_used { get; set; }
+        public long hours_used { get; set; }
         public string date_opened { get; set; }
         public string date_added { get; set; }
     }
